starting on WaterCup
I created the first "drop"of water for the project titled "WaterCup." This is a very simple game that will have a timer and a game manager. The inspiration comes from the water jug at my job. There's always something: people leave the tap open and wet the floor, it's always running out, we always need more cups, and it's also just an annoying repetitive thing. This is gonna be a simpler version of a small part of a game that I want to. work on in the future, which gives like cooking mama vibes but is about being a barista. But I digress.